BUG#71766 BUG#18280004: MYSQLFABRIC GROUP DESTROY COMMAND FAILURE
If you try to destroy a group when the group is used in the shard sub-system,
the following output is printed out:

-bash-4.1$ mysqlfabric group destroy my_group1
Procedure :
{ uuid        = 09ea87a2-8255-4856-bd84-d0be4fc25975,
  finished    = True,
  success     = False,
  return      = DatabaseError: ("Command (DELETE FROM groups WHERE group_id =
%s, ('my_group1',)) failed: 1451 (23000): Cannot delete or update a parent
row: a foreign key constraint fails (`fabric`.`shards`, CONSTRAINT
`fk_shards_group_id` FOREIGN KEY (`group_id`) REFERENCES `groups`
(`group_id`))", 1451),
  activities  =
}
 
The DatabaseError is not a user-friendly message. In this patch, we check
whether the group is used in the shard sub-system before trying to destroy
(i.e. remove) it. If it is in use, one of the following messages can be
printed out:

raise _errors.GroupError(
    "Cannot erase a group (%s) which is associated to a shard (%s)." % 
    (group_id, shard_id)
)

or

raise _errors.GroupError(
    "Cannot erase a group (%s) which is used as a global group in a "
    "shard definition (%s)." % (group_id, shard_mapping_id)
)