WL#7648: Decoupling provisioning from updates to state store
This patch introduces a new parameter in order to decouple provisioning from
updates to the state store. Any command that executes a provisioning step shall
provide the --update_only option so that users may choose whether they want to
execute the provisioning  steps or skip them. By default the option is false
meaning that the provisioning  shall be executed. If the option is set to true,
the command will only update the state store.

The --update_only parameter was added to the following commands:

. group.add(server, ..., update_only=False, ...)

. group.promote(group_id, slave_uuid=None, update_only=False, ...)

. group.demote(group_id, slave_uuid=None, update_only=False, ...)

. server.set_status(server, ..., update_only=False, ...)

. sharding.move(shard_id, ..., update_only=False, ...)

. sharding.split(shard_id, ..., update_only=False, ...)

We also took the opportunity to do the following changes in this patch:

. Removed the group.import_topology(...). It required to start servers
  with the following options: --report-host and --report-port. This was
  not very useful.

. Renamed group.check_group_availability(...) to group.health(...) and
  removed the "is_master" from the returned value. Users can check the
  same information in the "status" value.