Post-push fix for WL#7455.
"test_command" is failing in jenkins after WL#7455 because it creates a
remote command and invokes its dispatch method which sends a request
through the XMLRPC protocol. The XMLRPC interface though assumes that
the user/password are defined in the "protocol.xmlrpc" section in the
configuration. However, they are not defined and the execution fails.

In this patch, we define user/password in the "protocol.xmlrpc" section
before invonking the remote command.