BUG#72016 BUG#18403885 AUTOMATIC FAILOVER IS NOT HAPPENING WHEN GLOBAL GROUP MASTER IS FAILED
The shard sub-system creates a replication channel between a global group and
a shard group. Specifically, the master in the global group replicates its
changes to the master in the shard group.

Although there is this relationship, the high availability routines should
work independently in the sense that any high availability operation in the
groups should work regardless the existence of a valid master or not in the
related group(s). This was not happening though and an exception was raised
thus aborting the high availability operation(s).

We have fixed this behavior in the context of this patch. Now if there is
no valid master in a related group, an error message is printed out and
the operation continues.