BUG#18477189: MYSQLFABRIC PARAMETERS NEED TO BE PASSED TO CMDS WHEN AUTHENTICATION IS ENABLED
When credentials are enabled the command-line script is directly accessing
the backing store to authenticate a command. This is not supposed to happen
though and commands should contact Fabric and their components through a
remote procedure call (e.g. XML-RPC).

We fix this issue in this patch. Note though that user.* and role.* commands
were designed to run on the same machine that Fabric is running. So if one
tries to run these commands from a remote machine the appropriate
configuration must be defined so that they can direct access the backing
store.