BUG#72432: BUG#18458461: "MYSQLFABRIC MANAGE SETUP" RESULTS IN A LOST CONNECTION ERROR
The connection between Fabric and MySQL server has a timeout which might
be changed to a different value according to the environment where Fabric
is being deployed. If the timeout expires during the setup operation, the
setup does not undo its work and the MySQL server is left with incomplete
changes. And any attempt to run setup or teardown after that fails with
an error.

In this patch, we have kept the default timeout value and made the setup
behave atomically.