BUG#72149: BUG#18524482: EMPTY PASSWORD IN CONFIGURATION FILE LEADS TO ERRORS
The credential sub-system provides the means to authenticate users and
check if they have the necessary privileges to execute an operation.
In order to ease the MySQL Fabric administration, it is possible to
disable the credential sub-system or define a default user and password
in the configuration file which will be used whenever the 
mysqlfabricscript is executed. 
  
An empty user or password leads to unpredicatable behavior and crashes
though. It was possible to set up MySQL Fabric and skip the password
definition and mysqlfabric was starting with an empty password. This
patch fixes both problems. It changes the code to avoid empty passwords
or passwords with leading and trailing withe spaces.