BUG#72119: BUG#18454737 "GROUP ADD" HANGS WITH SERVER AS LOCALHOST:32274
The group.add or server.lookup_uuid commands were blocking if the server
was unresponsive or was not a MySQL instance. For example, the former
command would block if you tried to add the XML-RPC Server (e.g. localhost
:32274) to a group.

The first operation of adding a MySQL instance to group is establishing a
connection to it and retrieving its UUID. So to fix both issues, we have
introduced a timeout in the operation that checks/retrieves the UUID from
a MySQL instance.