BUG#72553: BUG#18712020: CONCURRENCY CONTROL IS NOT LOCKING THE RIGHT OBJECTS
A procedure in Fabric is scheduled according to the set of objects that are
provided to the scheduler by the procedure itself when it is about to start.
These objects are called lockable objects and are the group identifiers.

This means that update operations on the same group are serialized.

In order to ease the developer's life, the current implementation tries to
automatically find out the lockable objects for a given procedure by
analyzing the procedure's input parameters. Although this solution seems
fine, it is really hard to keep it updated.

Due to the several changes in the high-availability and sharding sub-systems,
this automatic routine is no longer valid. The consequence is that there may
be procedures trying to update the same group concurrently when they should
be serialized.

This patch provides a quick and temporary solution. It serializes all the
procedures. In the future, we are going to improve this though.