BUG#72606: BUG#17747197: MYSQL FABRIC: FAILOVER SCENARIO FAILED FOR A GROUP
Fabric uses threads to process XML-RPC requests (Session Threads),
update the state store (Executor Threads) and periodically check
servers' availability (Failure Detector Threads). Each thread gets
assigned its own database connection when it starts and clearly
the maximum number of database connections might not be enough to
serve all threads and some of them might fail.

Users must configure the number of threads judiciously as high number
of threads can not only exhaust the number of database connections
but might truly harm performance.

Threads are created when Fabric starts and when a group is activate
(i.e. a Failure Detector is started to monitor it). So before starting
Fabric or spawning a new Failure Detector, the maximum number of
available connections is checked and an exception is raised if it is
not possible to create a new connection to serve the new thread.