Post-push fix for WL#7426.
In python 2.6, the Filter class is an old-style class and does inherits
from object. This means that classes that inherit from the Filter class
cannot use super. We were using super in the MySQLFilter class and the
code was not compiling/running with python 2.6. To fix that, we have
used the old-style of calling a method from an inherited class.