BUG#72118: BUG#18454679: PROMOTING A SLAVE AFTER A RESET MASTER HANGS
"PROMOTE" after "RESET MASTER" is making Fabric hang due to apparently a
change in the replication's behavior. After the "RESET MASTER", the GTID
EXECUTED variable is cleaned and the "PROMOTE" operation, specially the
failover code, is fetching the GTID RETRIEVED variable and is waiting until
all retrieved transactions are reprocessed. However, apparently the relay
log(s) is not read from the beginning anymore and thus the failover
operation hangs.

To fix the problem, we have switched to coordinates and used the
Master_Log_File and Read_Master_Log_Pos reported by the "SHOW SLAVE STATUS"
to call "SELECT MASTER_POS_WAIT".