Post-push fix for WL#7426
WL#7426 creates a infra-structure to log statistics into tables and builds
upon the logging system. Users who want to write information to the
statistic table must set the "extra" parameter in the calls to the logging
with additional keys and values. If the logging level defined matches the
logging call, the information is written to the default handler (i.e. file
or stdout) and also the table.

However, we want a different behavior. Whenever the extra parameters are
specified, we want to write to the statistic table and only write to the
default handler when the logging level set by the user matches the logging
call. To implement this behavior, you set the logging level in the logger
and the statistic table handler to DEBUG, and the logging level in the
standard handler is set accordingly to the user's request.