WL#7749: Create a clone command for cloning servers
The clone command represents the pre-setup for provisioning a server in Fabric.
It represents the first set of steps that need to be performed before bringing a
server into the Fabric Farm.

The clone forms a smaller part of the bigger provisioning picture. In the
transition of the MySQL server between its various Fabric states, Clone
provides the set of steps required to move a MySQL Server from outside into
Fabric ecosystem.