Bug#18824565:SHARD MOVE AND SPLIT FAILS IF PASSWORD IS PROVIDED FOR CLIENT IN FABRIC.CFG
Using a fabric setup with a valid password in the fabric.cfg client section
was causing shard move and shard split to fail. This was because of erroneous
invocation of the mysql client command line.

This has been fixed by removing the misplaced mysql client parameter that
was being passed while calling the mysql client.