WL#7760: Command result interface to protocol classes
This patch implement a standardized interface between the commands and protocol
classes that is agnostic to the protocol being used. The interface is defined
as a command result class consisting of basic information and result sets.

In order to simplify implementation, the basic functions used internally,
generate_output_format and wait_for_procedures, are changed to read the
existing format from the functions and transform it to a result set. New
functions should use the command result interface, and some of the existing
functions have been changed to use the new interface.

The XML-RPC packet format have also been changed since the old format were
difficult to parse and also did not support extending.

In addition, the patch contain some changes to add MySQL version comments to
some constructions that are introduced in MySQL 5.6.

Some of the tests have been changed to handle the new packet format, but not
all tests have been converted yet.