WL#7523:WL#7539:Supporting DATETIME and STRING as sharding keys in Fabric.
It should be possible to define a RANGE based sharding scheme that uses a
DATETIME or a STRING column value as a sharding key.

The current implementation of RANGE sharding, supports sharding of tables on
sharding columns of the INTEGER type. But there are several other valid use
cases for RANGE based sharding using other datatypes.

This patch aims to add support for specifying DATETIME values and STRINGs in
the UTF8 format as sharding keys in the RANGE based sharding scheme.