Bug#18669231: SHARING PRUNE ISSUE WITH 20MN RECORDS
Prune was taking a very long time when the number of records that
needed to be pruned was a very large value (for e.g.
20MN records). This was caused by the DELETEs taking a very
long time when the data is very huge.

This can be avoided by performing the DELETEs in small bunches,
(i.e.) by setting a LIMIT on the number of records deleted at
any time.