BUG#19226319 - STATISTICS PROCEDURE CALL_ABORT ERROR
< mysqlfabric statistics procedure > of MySQL Fabric 1.5.1 doesn't return the 
correct value for call_abort for any procedure. The value returned stays at 0 
inspite of the procedures getting aborted. This issue is not present with the
earlier versions 1.4.3, 1.4.4.

This problem was cause by the WL#7760 which has changed how errors are reported.

Fixed the problem by checking whether there is an error reported in the object
CommandResult or not.