WL#7600: New interface using the MySQL Protocol
We add a new way to communicate with MySQL Fabric using
the MySQL Client-Server protocol. We have implemented
a TCP service which mimics the MySQL Server so all
MySQL connectors should be able to query for information
and executing Fabric commands. We call this MySQL-RPC.

Connecting with the MySQL-RPC service of Fabric is done
the usual way with username and password (currently only
mysql_native_password is supported). If the service runs
on the default port of Fabric:

  shell> mysql -uadmin -p --port 32274 --host example.com

Executing Fabric commands is done in a similar way as
executing stored routines, using the CALL statement. For
example:

  mysql> CALL dump.servers(patterns='%global');

The above command would return two resultsets. One with
generic Fabric information (UUID, TTL and a message).
The next resultset is the actual return of the command.

For a list of commands, use the mysqlfabric command line
tool, or send a query to the MySQL Fabric server:

  mysql> SELECT * FROM INFORMATION_SCHEMA.ROUTINES;

Note that the above SELECT will ignore column selections
and providing a WHERE-clause will not filter. Also, the
returned columns is a subset of the original table
found with a real MySQL Server.

To get help of a specific command, you can execute the
SHOW CREATE PROCEDURE statement. It will show the output
similar as what you would get using the mysqlfabric
command line tool:

  mysql> SHOW CREATE PROCEDURE dump.servers;

It is possible to get the output of the CALL statements
as JSON. You have to set the format variable before
executing CALL:

  mysql> SET format=json;
  mysql> CALL ..

An example usage, using the MySQL Connector/Python (note
the multi=True option):

  cur = cnx.cursor(dictionary=True)
  stmt = "CALL dump.servers(patterns=%s)"
  results = cur.execute(stmt, ('%global',), multi=True)
  for result in results:
    print result.fetchall()

The configuration of the MySQL-RPC interface is done
in the fabric.cfg file using the [protocol.mysql] section.
It is similar to the protocol.xmlprc section.

Credentials have been updated so a new 'admin' user is
created for this new protocol. It is thus possible that
the XML-RPC and MySQL-RPC interface have different
credentials.