BUG#19277230 - UPDATE THE REPOSITORY TO MAKE A STAND ALONE RELEASE
Fabric will no longer be a subproduct of Workbench. Thus any reference to
the Workbench must be removed and the README file must be updated with
appropriate license information.
      
There was no reference to the Workbench in the Fabric's repository. So to
fulfill this requirement, we have kept the original README file in the
root directory untouched and created the following files: 
      
  . release/community/LICENSE.txt
  . release/community/README