BUG#19289261 - DATA IN THE DUMP.SHARDING_INFORMATION IS MISPLACED
The dump.sharding_information should return a result set with the following 
sequence of columns: 

 1. 'schema_name' 
 2. 'table_name' 
 3. 'column_name' 
 4. 'lower_bound' 
 5. 'shard_id' 
 6. 'type_name' 
 7. 'group_id' 
 8. 'global_group' 

However, the following is being returned: 

 1. 'schema_name' 
 2. 'table_name' 
 3. 'column_name' 
 4. 'lower_bound' 
 5. 'shard_id' 
 6. 'group_id' 
 7. 'global_group_id' 
 8. 'type_name' 

Note that columns 6, 7 and 8 are not in the expected position. Besides, 
global_group_id should be renamed to global_group.

In this patch, we fixed this issue and took also the opportunity to
fix some test cases that were failing after BUG#19226122.