Bug#19450843: PRUNE DOES NOT WORK IF TABLE DOES NOT EXIST
During a Shard Prune or a Shard Split, if the shard mapping exists for the
table exists, but the table itself does not exist in the server, both the
operations fail.

Fabric will focus on the metadata it contains and will not verify the data
present in the database servers. Thus we will throw an error if the shard
mapping is not defined, but we will not verify whether the table exists on the
server or not.

The patch catches the DatabaseError that is thrown when the DELETE query is
fired on the non-existent tables and ensures that it is not passed to the user.