BUG#19172889: BUG#73233: CMDS THAT ACCEPT BOTH A UUID AND ADDRESS MIGHT HANG IF A WRONG ADDRESS IS GIVEN
Commands accept either a UUID or an address (i.e. host:port) to
specify a server. The code is ignoring though the fact that a
wrong address might be provided and in consequence the execution
might hang similar to what happened in BUG#72119.

In this patch, we do a minor refactoring and check whether the
address is valid or not before using it.

If the server belongs to a group, however, the address is checked
against the information stored in the state store.

If the server is not part of a group, Fabric will try to access
the server until a timeout is reached. The default value for this
timeout is defined through the option "unreachable_timeout" in
the "servers" section in the confugration file and affets the
following commands:

 . server lookup_uuid address [--timeout=NONE] ...
 . server clone group_id destn_address [--timeout=NONE] ...
 . group add group_id address  [--timeout=NONE] ...