BUG#72931: BUG#18991971: MYSQLFABRIC GROUP DESTROY --FORCE DOES NOT WORK, ERROR IN _DESTROY_GROUP.
The --force option was introduced to allow users to delete a group even
when there are servers associate to it provided those servers are not
being used by the sharding sub-system. Apparently though, this is causing
some confusion as users expect to have the group deleted no matter what
the dependencies are.

To turn things more complicated, some of the dependencies were not
considered and an attempt to delete a group when it was not being used by
the sharding sub-system could raise an unfriendly error message.

Making the command (group destroy --force) delete all servers and
associated sharding information is risky though as users may delete
things by mistake and may harm both the HA and sharding sub-system.

So taking this into consideration, we have decided to remove the --force
option and make sure that friendly messages are always printed out.