Bug#73218: mysqlfabric hash sharding table shard_ranges lower_bound maybe 32 byte length
While moving to a sharded from a unsharded setup in
Hash based sharding, splitting of the first unsharded
server can result in generating a lower_bound value,
than cannot be stored in a VARBINARY(16) column.

The problem was being ignored silently when the sql_mode
was not set to STRICT_ALL_TABLES. This can be a problem
since a truncated lower_bound value is a deviation from
the hash ring.

The issue was happening because an attempt was being made
to covert the output of the MD5 function (which is already
hexadecimal) into hexadecimal again by HEXing it.

The patch fixes the issue by avoiding the unnecessary HEX.