Bug#19427027: RESHARDING CAN CAUSE INCONSISTENCY DURING HIGH LOADS
During a shard move or a split, the source MySQL group
is made read only before synchronizing the destination
and changing the sharding metadata to match the result
of the resharding. There is a danger that transactions
that arrived during the period for which the group is
read only, can get unblocked, and write wrong information
into the source group. This can result in inconsistent
information.

The patch ensures consistency by doing the following, during
the re-sharding algorithms

- Ensuring the source and the destination groups are read-only
  post the synchronization
- Switch the shard mappings
- Wait till TTL expires for the shards to refresh their information
- Kill Processes on the source machine
- Change the permissions back to read-write