BUG#73873 : BUG#19594070: LOCKING ERROR IN THE FAILURE DETECTOR
The failure detector triggers a failover operation that should
not be allowed to concurrently run with other operations that
update the same group. In order to guarantee this, Fabric relies
on a locking mechanism that prevents concurrent updates to a
group.
      
Locks are acquired before an execution starts and are released at
its completion. Currently, Fabric serializes all update operations
due to problems in the mechanism that determines which groups should
be locked. One single lockable object is used: set(['lock']). See
BUG#72553: BUG#18712020 for further details.

However, the failure detector was specifying a set of groups as
lockable objects and a failover operation could potentially run
while a group was being updated and lead to unpredictable results.
      
In this patch, we make the failure detector use the following
lockable objects: set(['lock']).