Bug#74339: Bug#19807933: FABRIC SHOULD CHECK FOR SUPPORTED CONNPY VERSION
The expected version of Connector/Python is stored in mysql.fabric package but
the comparison with the expected and current version is done using string
comparison, which fails if the current version is '2.0a1' and the required
version is '2.0.0' because '2.0a1' < '2.0.0' is false despite the fact that
'2.0a1' is before '2.0.0'. In addition, the check is not executed prior to
importing mysql.connector, so an exception is thrown if an attempt is made to
import symbols from mysql.connector that do not exist in earlier version.

This patch fixes the issue by using distutils.version.LooseVersion to compare
the versions and move the dependency check to before importing any mysql.fabric
sub-packages. We do not use distutils.version.StrictVersion since non-strict
versions (e.g., '1.2.3rc1') might temporarily be used and StrictVersion throws
an exception in this case.