BUG#74255: BUG#19774543: MYSQLSERVER OBJECT CREATES CONNECTIONS WITH "USE_UNICODE=FALSE"
MySQLServer object creates connections with "use_unicode=False" and as
a consequence some statements return fields in bytearray format. The
code though is expecting strings or unicode fields and the unexpected
field data types might cause problems to several modules such as
replication and high availability.
  
This happens because the routine that converts MySQL data types to
python data types has the following code:
  
    if isinstance(value, (bytes, bytearray)) and self.use_unicode:
        return value.decode(self.charset)
  
In this patch, we removed the "use_unicode=False" and made the following
changes:
  
  - Removed "raw": False, since that is the default
  - Changed from list to tuples "params": (param_1, param_2, ...) for
    constant inputs
  - Handled pylint warning "params": (param_1, )
  - Removed "fetch" : True, since that is the default
  - Removed some unused functions: get_master_slaves(), user_databases()