BUG#74555: BUG#19949241: MYSQL FABRIC HANGS ON NETWORK OUT ON A MASTER OR SLAVE MySQL
The problem here is that any function (e.g. built-in failure detector) that
tracks whether a server is unreachable or not should kill all connections to
the sever when the it is considered faulty thus indirectly aborting/unblocking
any command accessing the server.

This is not happening though and commands are blocked until the TCP-IP
channel times out and this might take hours.

To fix this issue, the built-in failure detector and the "threat group
command" will kill any connection to a faulty server thus releasing commands
that might be blocked.

Note though that the solution does not rely on any explicit socket/connector
timeout as there might be Fabric operations that can take hours. In other
words, Fabric will kill only those connections to servers that have failed
when it finds out about the failures either through the built-in failure
detector or the "threat group command".

In particular, the built-in failure detector will always check whether there
are new attempts to connect to a server that has failed and will kill these
new connections. However, if the built-in failure detector is not running, a
new command will block until the "socket.connect" times out. This timeout is
usually shorter (i.e. within seconds) and could be reduced without any harm
to the aforementioned long running operations.

Adding a server or reintegrating it into a group after a failure is done in
two steps now. In the addition, the server is added into a group, put into a
configuring state and in the next step its replication is set up and then
it is moved to a secondary state. In the reintegration, a faulty server is
moved to a configuring state and after its replication is set up.

This was done so the buit-in failure detector could monitor these servers
and didn't try to automatically kill any connection to it.

In this patch, we fixed the reported issue as previously described and took
the opportunity to make the following changes:

  . ConnectionPool is renamed to ConnectionManager.
  . ConnectionManager's method accepts a server instead of a UUID.
  . ConnectionManager creates new connections as well.
  . The server_utils.py is renamed to connection.py.
  . The ConnectionManager is moved into connection.py.
  . The "free" functions in connection.py are moved into a Connection class.
  . A timeout option was added to the "group health" command.
  . Minor refactoring.