BUG#73968: BUG#19644057: MYSQLFABRIC EVENT TRIGGER DOES NOT HANDLE PARAMETERS CORRECTLY
The command-line "event trigger" does not handle parameters correctly:

  class Trigger(Command):
    group_name = "event"
    command_name = "trigger"

  def execute(self, event, locks, *args, **kwargs):
    ....

The problem was exposed by the code pushed to release 1.4.2, in the context
of WL#7528 ("Adding support for optional arguments in Fabric"). After this
patch, MySQL Fabric stopped recognizing args and kwargs as parameters.

Although the args and kwargs were recognized as parameters before the
aforementioned WL, the command did not really had support to an
unbounded number of parameters. The problem stemmed from the fact that
the XML-RPC does not have support to the non-keyworded and keyworded
parameters.

So if you have tried to fire an event that required some parameters you
would get an error. For example,

  $ mysqlfabric -config event trigger EVENT my_group \
    8b1e25db-c8f4-43a2-bd07-79335753f693
  Usage:  event trigger event locks

  mysqlfabric: error: Wrong number of parameters were provided for command
  'event trigger'.

In this patch, we make the 'event trigger' command accept non-keyworded and
keyworded parameters as follows:

  $ mysqlfabric help event trigger
  event trigger event  [--locks=NONE] [--args=NONE] [--kwargs=NONE]  

  Trigger an event.

For example,

  $ mysqlfabric event --locks group --args arg_1 --args arg_2 --kwargs \
    key_1=value_1 --kwargs key_2=value_2