WL#7860: Prevent wrong inserts/updates into shards
The worklog shall aim to prevent users from firing inserts or updates that can
result in the sharding column taking on values that fall outside the range of
valid values for a shard.

The current worklog aims to implement shard range checks using TRIGGERs. The
goal is only to ensure that a row is not accidentally "lost" by storing it in a
shard outside the shard range. Thus we will worry specically about INSERT and
UPDATE and aim to create triggers for the same.

The triggers are created on the tables that are registered with Fabric. The
user can chose while setting up sharding the tables that need triggers created
on them for checking the key value range in the shard column.

Servers now contain the valid range of values stored in them in the
mysql_shard_range table in the mysql_fabric database.

The triggers work by verifying that the inserts and updates keep to the valid
range of values that are stored in the metadata tables on servers registered in
Fabric.