Post-push fix for BUG#74555: BUG#19949241.
Catching multiple exceptions in one single line was broken in some
places as the old style was being used thus unexpectedly changing
some objects' values and causing strange errors. For example, the
following code assigned a ServerError instance into DatabaseError
class when the ServerError was caught:

  try:
    ...
  except _errors.ServerError, _errors.DatabaseError:
    ...

We have fixed that by putting the exceptions into parenthesis.

Besides the failover operation was not working properly when there
was a server with the CONFIGURING status. These servers are not
valid candidates but were chosen thus making the failover fail.