BUG#20766081 FABRIC DAEMON HANGS ON GROUP COMMANDS FOR SOME UNKNOWN REASON
MySQL Fabric is hanging due to a bug in connector python that has been
fixed in the 2.0.4 release. See BUG#74933 for further details.

MySQL Fabric cannot perform any operation when connector python 2.1.1 is
installed. See BUG#76579 for further details.

In the context of BUG#20766081, we replaced the dependency on the 2.0.1
connector python release for the 2.0.4 or 2.1.2 releases.

Why two different releases?

Currently, connector python has two parallel releases: 2.0.X without
c-extensions and 2.1.X with C-extensions. Eventually, the 2.0.X will be
deprecated but for now we have to keep track of both.