BUG#21042726 : MYSQL FABRIC SPINS IN RECVFROM AFTER MYSQL RPC CONNECTION CLOSES
In the MySQL RPC protocol, the MySQL Fabric process spins in the "recv"
after connector closes the connection causing excessive (100%) CPU use.

To fix the bug, we have changed the code to detect that an empty chunk
is returned from "recv" meaning that nothing else will be transferred
from the client and that the socket has been closed. We also have taken
the opportunity to fix the case where headers are fragmented avoiding
that the process hangs trying to read 4 bytes.