BUG#21043598 : MYSQL FABRIC ALLOWS SERVERS WITHOUT PORTS IN SERVER_ADDRESS COL
The "group.add" command accepts a server address without a port and
automatically assumes the default 3306, but it stores the original
information provided by the user. However, not all commands check
whether the address has the format host:port. This behavior might
cause some issues if an application fetches the address and blindly
tries to use it.

In this patch, we force that an address is always converted to the
host:port format before being stored into the state store and a
warning message is printed out if the address does not contain the
port and the default port is used.