BUG#20230459: NEED SIGNAL HANDLER TO DUMP STACK WHEN SENT SIGUSR1
(This patch was manually backported from the master branch)

In this patch, when the SIGUSR1 signal is sent to the MySQL Fabric
process, the stack trace of all threads are sent to the current
logger, if there is one, or to the 'stderr' stream otherwise. To
decide whether the the logger or the 'stderr' stream should be
used, the handler function is wrapped within another function and
exploits closures in python.

The patch also defines an empty handler for the SIGINT signal
avoiding that someone kills the MySQL Fabric process by mistake
after pressing ctrl + c. To stop the MySQL Fabric process, the
following alternatives may be used:

. executing 'mysqlfabric manage stop'.

. sending the SIGKILL signal to it.

In this patch, we also took the opportunity to fix two additional
bugs.

First, the 'daemonize' option, which is used to start the MySQL
Fabric process as a daemon, was not working properly. The function
that builds the list of options extracts them from the 'execute'
method provided it is defined. Otherwise it extracts them from the
dispatch method. However, to avoid PyLint warnings, an 'execute'
method was introduce in the base class 'Command' and the dispatch
function where the daemonize option is currently defined was never
checked:

  commit bc53d9d59dfaaf140d2dbcdae3e5625fb9c661f9
  Author: Mats Kindahl mats.kindahl@gmail.com
  Date: Thu Mar 26 12:47:33 2015 +0100

  Fixing some high and medium PyLint warnings.

To fix the issue, we checked whether the 'execute' method was defined
only by the base class 'Command' or was overridden by a sub-class. In
the former case, arguments are extracted from the dispatch method to
build the options and help texts. In the latter case, they are
extracted from the execute method.

Second, the stderr was not redirected when the MySQL Fabric process
was daemonized. So the process would die if someone tried to dump a
stack trace. To fix the first issue, we redirected the 'stderr' to
'os.devnull' (i.e. '/dev/null' for POSIX).