Bug#21214069: OPTIONAL PARAMETERS IN DISPATCH METHODS ARE NOT PARSED
Description
-----------

Dispatch methods of the command class created with more than one
compulsory arguments would throw an error saying that the number of
arguments supplied is not correct.

Fix
---

The fire_command method in the mysqlfabric module has a check for
verifying the number of arguments required by a command and the number
of parameters actually passed. This logic was working with the execute
methods all the time after the default execute method implementation in
the Command class. The logic in this method has been updated to use the
dispatch methods when available.