BUG#77463: BUG#21307183: FABRIC NODE FAILOVER FAILED AFTER SHUTTING DOWN PRIMARY NODE
MySQL Fabric has been started properly but the fabric agent
thought that there has been some issue because it could not
find a MySQL Fabric process with the same pid stored in the
pid-file.

Apparently, this was happening because the pid-file had been
created by the parent process and the child process, which
was created after a fork, did not have the chance to execute
the MySQL Fabric program.

To avoid the issue, we check if there is a valid MySQL Fabric
process a few times before giving up.