BUG#20697968: MYSQLFABRIC CLIENT FREQUENTLY ERRORS WITH 'CONNECTION RESET BY PEER'
MySQL Fabric implements two RPC protocols, MySQL and XML namely.
The former spawns a new thread to handle an incoming request
while the latter has a pre-defined number of threads to handle
incoming requests.

Having an admission control system is key to avoiding overloading
MySQL Fabric and the state store (i.e. MySQL Server). Note that
each thread needs a connection to the state store and an admission
control system avoids reaching the maximum number of database
connections in the state store.

The XML-RPC protocol implements this admission control system which
assigns a request to an idle thread. If there is no idle thread,
however, the thread responsible for accepting incoming requests
(i.e. the one which calls the socket.accept()) blocks until a
thread becomes idle.

This means that any new request will be put in the TCP backlog which
is 5 by default. So if the number of threads is not carefully
configured, clients may get the "CONNECTION RESET BY PEER" or
"CONNECTION TIMED OUT" errors.

This behavior though was causing some confusion and for that reason
we have decided to change it. In this bug, incoming request are
immediately put into a queue where threads will pull from. Note
though that the "CONNECTION RESET BY PEER" might happen if there
are network issues as well.